separated screen logic from app.js
